+++ dart/collision/bullet/BulletCollisionDetector.cpp
@@ -594,6 +594,14 @@ void reportContacts(
     for (auto j = 0; j < numContacts; ++j)
     {
       const auto& cp = contactManifold->getContactPoint(j);
+
+      if (cp.m_normalWorldOnB.length2() < Contact::getNormalEpsilonSquared())
+      {
+        // Skip this contact. This is because we assume that a contact with
+        // zero-length normal is invalid.
+        continue;
+      }
+
       result.addContact(convertContact(cp, collObj0, collObj1));
 
       // No need to check further collisions