+++ dart/gui/OpenGLRenderInterface.hpp
@@ -93,7 +93,7 @@ public:
     void drawSoftMesh(const aiMesh* mesh) override;
     void drawList(GLuint index) override;
     void drawLineSegments(const std::vector<Eigen::Vector3d>& _vertices,
-                                  const dart::aligned_vector<Eigen::Vector2i>& _connections) override;
+                          const common::aligned_vector<Eigen::Vector2i>& _connections) override;
 
     void setPenColor(const Eigen::Vector4d& _col) override;
     void setPenColor(const Eigen::Vector3d& _col) override;