+++ dart/math/MathTypes.hpp
@@ -38,16 +38,13 @@
 
 #include <Eigen/Dense>
 
-#include "dart/common/Deprecated.hpp"
-#include "dart/common/Memory.hpp"
-
 //------------------------------------------------------------------------------
 // Types
 //------------------------------------------------------------------------------
 namespace Eigen {
 
-typedef Matrix<double, 6, 1> Vector6d;
-typedef Matrix<double, 6, 6> Matrix6d;
+using Vector6d = Matrix<double, 6, 1>;
+using Matrix6d = Matrix<double, 6, 6>;
 
 inline Vector6d compose(const Eigen::Vector3d& _angular,
                         const Eigen::Vector3d& _linear)
@@ -57,48 +54,15 @@ inline Vector6d compose(const Eigen::Vector3d& _angular,
   return composition;
 }
 
-typedef std::vector<Eigen::Vector3d> EIGEN_V_VEC3D;
-typedef std::vector<std::vector<Eigen::Vector3d > > EIGEN_VV_VEC3D;
-
 } // namespace Eigen
 
 namespace dart {
-
-#if EIGEN_VERSION_AT_LEAST(3,2,1) && EIGEN_VERSION_AT_MOST(3,2,8)
-
-template <typename _Tp>
-using aligned_vector = std::vector<_Tp,
-    dart::common::detail::aligned_allocator_cpp11<_Tp>>;
-
-template <typename _Key, typename _Tp, typename _Compare = std::less<_Key>>
-using aligned_map = std::map<_Key, _Tp, _Compare,
-    dart::common::detail::aligned_allocator_cpp11<std::pair<const _Key, _Tp>>>;
-
-#else
-
-template <typename _Tp>
-using aligned_vector = std::vector<_Tp, Eigen::aligned_allocator<_Tp>>;
-
-template <typename _Key, typename _Tp, typename _Compare = std::less<_Key>>
-using aligned_map = std::map<_Key, _Tp, _Compare,
-    Eigen::aligned_allocator<std::pair<const _Key, _Tp>>>;
-
-#endif
-
-template <typename _Tp, typename... _Args>
-DART_DEPRECATED(6.2)
-std::shared_ptr<_Tp> make_aligned_shared(_Args&&... __args)
-{
-  return ::dart::common::make_aligned_shared<_Tp, _Args...>(
-        std::forward<_Args>(__args)...);
-}
-
 namespace math {
 
-typedef Eigen::Matrix6d Inertia;
-typedef Eigen::Matrix<double, 3, Eigen::Dynamic> LinearJacobian;
-typedef Eigen::Matrix<double, 3, Eigen::Dynamic> AngularJacobian;
-typedef Eigen::Matrix<double, 6, Eigen::Dynamic> Jacobian;
+using Inertia = Eigen::Matrix6d;
+using LinearJacobian = Eigen::Matrix<double, 3, Eigen::Dynamic>;
+using AngularJacobian = Eigen::Matrix<double, 3, Eigen::Dynamic>;
+using Jacobian = Eigen::Matrix<double, 6, Eigen::Dynamic>;
 
 }  // namespace math
 }  // namespace dart