+++ CHANGELOG.md
@@ -14,13 +14,22 @@
 
 ### DART 6.4.0 (201X-XX-XX)
 
-* Kinematics
+* Kinematics/Dynamics
 
   * Added IkFast support as analytic IK solver: [#887](https://github.com/dartsim/dart/pull/887)
+  * Fixed NaN values caused by zero-length normals in ContactConstraint: [#881](https://github.com/dartsim/dart/pull/881)
+
+* Collision detection
+  
+  * Added FCL 0.6 support (backport of #873): [#936](https://github.com/dartsim/dart/pull/936)
 
 * License
 
   * Added Personal Robotics Lab and Open Source Robotics Foundation as contributors: [#929](https://github.com/dartsim/dart/pull/929)
+  
+* Misc
+
+  * Suppressed warnings: [#937](https://github.com/dartsim/dart/pull/937)
 
 ### [DART 6.3.0 (2017-10-04)](https://github.com/dartsim/dart/milestone/36?closed=1)